backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.11.4
python-dotenv==1.0.1
google-generativeai==0.3.2 
mistralai==1.7.0 